project presentation
The presentation explores a case study of a territory in Shlisselburg that used to be an agricultural area. Now, this territory is being developed, and it is essential to determine the appropriate land use for this area. We used various instruments to compare the results and select the most suitable option. The presentation includes pictures of the different stages of development, along with brief descriptions and conclusions.